+++ examples/complete/main.tf
@@ -19,6 +19,14 @@ resource "azurerm_resource_group" "this" {
   tags = local.tags
 }
 
+module "log_analytics" {
+  source = "github.com/equinor/terraform-azurerm-log-analytics?ref=v1.4.0"
+
+  workspace_name      = "log-${random_id.this.hex}"
+  resource_group_name = azurerm_resource_group.this.name
+  location            = azurerm_resource_group.this.location
+}
+
 module "network_hub" {
   # source = "github.com/equinor/terraform-azurerm-network?ref=v0.0.0"
   source = "../.."
@@ -158,6 +166,18 @@ resource "azurerm_public_ip" "example" {
   location            = azurerm_resource_group.this.location
   sku                 = "Standard"
   allocation_method   = "Static"
+}
+
+module "public_ip" {
+  # source = "github.com/equinor/terraform-azurerm-network//modules/public-ip?ref=v0.0.0"
+  source = "../../modules/public-ip"
+
+  address_name               = "pip-${random_id.this.hex}"
+  resource_group_name        = azurerm_resource_group.this.name
+  location                   = azurerm_resource_group.this.location
+  log_analytics_workspace_id = module.log_analytics.workspace_id
+  sku                        = "Standard"
+  allocation_method          = "Static"
 
   tags = local.tags
 }
@@ -171,8 +191,8 @@ module "nat" {
   location            = azurerm_resource_group.this.location
 
   public_ip_associations = {
-    "this" = {
-      public_ip_address_id = azurerm_public_ip.example.id
+    "example" = {
+      public_ip_address_id = module.public_ip.address_id
     }
   }